medical_trustworthy_agent/requirements.txt
# --- Core ---
streamlit
python-dotenv

# --- LangChain Core ---
langchain
langchain-community
langchain-core

# --- Vector DB (FAISS) ---
faiss-cpu

# --- Embeddings ---
sentence-transformers
transformers
torch

# --- Optional utilities ---
pypdf
numpy
scikit-learn
